content/media/webaudio/test/test_audioBufferSourceNode.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test AudioBufferSourceNode</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="webaudio.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

var gTest = {
  length: 4096,
  createGraph: function(context) {
    var buffer = context.createBuffer(1, 2048, context.sampleRate);
    for (var i = 0; i < 2048; ++i) {
      buffer.getChannelData(0)[i] = Math.sin(440 * 2 * Math.PI * i / context.sampleRate);
    }

    var source = context.createBufferSource();
    ok("gain" in source, "AudioBufferSourceNode.gain must exist");
    is(source.gain.value, 1, "AudioBufferSourceNode.gain's default value must be 1");

    var sp = context.createScriptProcessor(2048);
    source.start(0);
    source.buffer = buffer;
    return source;
  },
  createExpectedBuffers: function(context) {
    var buffers = [];
    var buffer = context.createBuffer(2, 2048, context.sampleRate);
    for (var i = 0; i < 2048; ++i) {
      buffer.getChannelData(0)[i] = Math.sin(440 * 2 * Math.PI * i / context.sampleRate);
      buffer.getChannelData(1)[i] = buffer.getChannelData(0)[i];
    }
    buffers.push(buffer);
    buffers.push(getEmptyBuffer(context, 2048));
    return buffers;
  },
};

runTest();

</script>
</pre>
</body>
</html>
